Modularize shared interface assets
