+++ src/templates/404.html
@@ -1,27 +1,18 @@
-<!DOCTYPE html>
-<html lang="en">
+{% extends "base.html" %}
 
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>404 - Page Not Found</title>
-  <link rel="stylesheet" href="/static/css/output.css?v=amber-ui">
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Archivo:wght@400;500;600;700;800&family=Bricolage+Grotesque:wght@400;500;600;700;800&family=IBM+Plex+Mono:wght@400;500;600;700&family=Public+Sans:wght@400;500;600;700;800&display=swap" rel="stylesheet">
-</head>
+{% block title %}404 - Page Not Found{% endblock %}
+{% block sentry %}{% endblock %}
+{% block body_class %}app-page status-page{% endblock %}
 
-<body class="app-page status-page">
-  <main class="app-shell status-shell">
+{% block body %}
+<main class="app-shell status-shell">
     <section class="app-card status-card">
-      <p class="eyebrow">404</p>
-      <h1>Page Not Found</h1>
-      <p class="subtitle">The page you are looking for does not exist or the URL might be incorrect.</p>
-      <footer class="card-footer">
-        <p><strong>Need help?</strong> Please contact Raj on Teams.</p>
-      </footer>
+        <p class="eyebrow">404</p>
+        <h1>Page Not Found</h1>
+        <p class="subtitle">The page you are looking for does not exist or the URL might be incorrect.</p>
+        <footer class="card-footer">
+            <p><strong>Need help?</strong> Please contact Raj on Teams.</p>
+        </footer>
     </section>
-  </main>
-</body>
-
-</html>
+</main>
+{% endblock %}
