--- src/templates/_sentry.html
@@ -1,14 +1,0 @@
-<script
-  src="https://js.sentry-cdn.com/b1cf415940d1bc459daa1127317e541b.min.js"
-  crossorigin="anonymous"
-></script>
-<script>
-  Sentry.onLoad(function () {
-    Sentry.init({
-      integrations: [Sentry.replayIntegration()],
-      tracesSampleRate: 1.0,
-      replaysSessionSampleRate: 0.1,
-      replaysOnErrorSampleRate: 1.0,
-    });
-  });
-</script>
