+++ src/templates/error.html
@@ -1,28 +1,17 @@
-<!DOCTYPE html>
-<html lang="en">
+{% extends "base.html" %}
 
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Server Error</title>
-  <link rel="stylesheet" href="/static/css/output.css?v=amber-ui">
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Archivo:wght@400;500;600;700;800&family=Bricolage+Grotesque:wght@400;500;600;700;800&family=IBM+Plex+Mono:wght@400;500;600;700&family=Public+Sans:wght@400;500;600;700;800&display=swap" rel="stylesheet">
-  {% include "_sentry.html" %}
-</head>
+{% block title %}Server Error{% endblock %}
+{% block body_class %}app-page status-page{% endblock %}
 
-<body class="app-page status-page">
-  <main class="app-shell status-shell">
+{% block body %}
+<main class="app-shell status-shell">
     <section class="app-card status-card">
-      <p class="eyebrow">System Alert</p>
-      <h1>Internal Server Error</h1>
-      <p class="subtitle">Something went wrong on our end. We're investigating the issue.</p>
-      <footer class="card-footer">
-        <p><strong>Need help?</strong> Please contact Raj on Teams.</p>
-      </footer>
+        <p class="eyebrow">System Alert</p>
+        <h1>Internal Server Error</h1>
+        <p class="subtitle">Something went wrong on our end. We're investigating the issue.</p>
+        <footer class="card-footer">
+            <p><strong>Need help?</strong> Please contact Raj on Teams.</p>
+        </footer>
     </section>
-  </main>
-</body>
-
-</html>
+</main>
+{% endblock %}
